requirements.txt
streamlit
openai>=1.0.0
PyPDF2
python-dotenv
google-genai
requests
groq